feat: wasm add fn working in react component

+++ ppeg-wasm/src/lib.rs
@@ -1,7 +1,7 @@
 use wasm_bindgen::prelude::wasm_bindgen;
 
 #[wasm_bindgen]
-pub fn add(left: u64, right: u64) -> u64 {
+pub fn add(left: i32, right: i32) -> i32 {
   left + right
 }
